+++ applications/external/resistors/src/scene_edit.c
@@ -8,16 +8,16 @@
 #include <applications/services/gui/modules/widget_elements/widget_element.h>
 #include <applications/services/gui/view.h>
 
-const int values_left = 64;
-const int rows_tops[] = {0, 9, 18};
-const int footer_top = 55;
+const uint8_t values_left = 64;
+const uint8_t rows_tops[] = {0, 9, 18};
+const uint8_t footer_top = 55;
 
-const int bands_top = 30;
-const int bands_lefts[] = {18, 34, 50, 66, 86, 102};
-const int band_indices[6][6] =
+const uint8_t bands_top = 30;
+const uint8_t bands_lefts[] = {18, 34, 50, 66, 86, 102};
+const uint8_t band_indices[6][6] =
     {{}, {}, {0, 1, 2}, {0, 1, 2, 5}, {0, 1, 2, 3, 5}, {0, 1, 2, 3, 4, 5}};
-const int band_w = 8;
-const int band_h = 22;
+const uint8_t band_w = 8;
+const uint8_t band_h = 22;
 
 void resistors_edit_view_redraw_widget(App* app) {
     widget_reset(app->widget);
@@ -47,14 +47,15 @@ void resistors_edit_view_redraw_widget(App* app) {
 
     // render band indicator
     if(app->state->edit_selection < app->state->resistor_type) {
-        int band_index = band_indices[app->state->resistor_type - 1][app->state->edit_selection];
+        uint8_t band_index =
+            band_indices[app->state->resistor_type - 1][app->state->edit_selection];
         widget_add_icon_element(app->widget, bands_lefts[band_index], bands_top, &I_box_8x22);
     }
 
     // render band colour descriptors (short)
-    for(int i = 0; i < app->state->resistor_type; i++) {
-        int description_index = band_indices[app->state->resistor_type - 1][i];
-        int description_left = bands_lefts[description_index];
+    for(uint_fast8_t i = 0; i < app->state->resistor_type; i++) {
+        uint8_t description_index = band_indices[app->state->resistor_type - 1][i];
+        uint8_t description_left = bands_lefts[description_index];
         widget_add_string_element(
             app->widget,
             description_left,
